Introduces function to set background color
Up until this point, I had to manually tweak numbers in the .c file to make sure, that after changing colorscheme, I'm going to get a proper border color around the edit are. Now there's a function which can be called from inside the Vim script and it can dynamically set the background color, which is nice!